# Add the feature to `JournalStorage` of Redis backend to resume from a snapshot