--- a/optuna/storages/_journal/base.py
+++ b/optuna/storages/_journal/base.py
@@ -1,7 +1,9 @@
 import abc
 from typing import Any
+from typing import Callable
 from typing import Dict
 from typing import List
+from typing import Optional
 
 
 class BaseJournalLogStorage(abc.ABC):
@@ -40,4 +42,39 @@ class BaseJournalLogStorage(abc.ABC):
                 A list that contains json-serializable logs.
         """
 
+        raise NotImplementedError
+
+
+class SnapshotRestoreError(Exception):
+    """Exception for BaseJournalLogSnapshot."""
+    pass
+
+
+class BaseJournalLogSnapshot(abc.ABC):
+    """Optional base class for Journal storages.
+
+    Storage classes implementing this base class may work faster when
+    constructing the internal state from the large amount of logs.
+    """
+    @abc.abstractmethod
+    def save_snapshot(self, snapshot: bytes) -> None:
+        """Save snapshot to the backend.
+
+        Args:
+            snapshot: A serialized snapshot (bytes)
+        """
+        raise NotImplementedError
+
+    @abc.abstractmethod
+    def load_snapshot(self, loader: Callable[[bytes], None]) -> None:
+        """Load snapshot from the backend.
+
+        Args:
+            loader: A callback function which accept one positional argument.
+                This callback is supposed to be called inside the
+                :meth:`~optuna.storages._journal.base.BaseJournalLogSnapshot.load_snapshot`
+                method.
+                This callback may raise :class:`optuna.storages._journal.base.SnapshotRestoreError`
+                when a serialized snapshot object is invalid or broken.
+        """
         raise NotImplementedError