--- a/optuna/storages/_journal/base.py
+++ b/optuna/storages/_journal/base.py
@@ -49,14 +49,9 @@ class BaseJournalLogSnapshot(abc.ABC):
 
     Storage classes implementing this base class may work faster when
     constructing the internal state from the large amount of logs.
-
-    Args:
-        snapshot_interval: An interval of trials and studies to save a snapshot.
     """
 
-    def __init__(self, snapshot_interval: int) -> None:
-
-        self.snapshot_interval = snapshot_interval
+    snapshot_interval = 100
 
     @abc.abstractmethod
     def save_snapshot(self, snapshot: bytes) -> None: