--- a/optuna/storages/_journal/redis.py
+++ b/optuna/storages/_journal/redis.py
@@ -9,7 +9,6 @@ from optuna._experimental import experimental_class
 from optuna._imports import try_import
 from optuna.storages._journal.base import BaseJournalLogSnapshot
 from optuna.storages._journal.base import BaseJournalLogStorage
-from optuna.storages._journal.base import SnapshotRestoreError
 
 
 with try_import() as _imports: