--- a/optuna/storages/_journal/storage.py
+++ b/optuna/storages/_journal/storage.py
@@ -31,6 +31,8 @@ from optuna.trial import TrialState
 _logger = optuna.logging.get_logger(__name__)
 
 NOT_FOUND_MSG = "Record does not exist."
+# A heuristic interval number to dump snapshots
+SNAPSHOT_INTERVAL = 100
 
 
 class JournalOperation(enum.IntEnum):
@@ -155,7 +157,7 @@ class JournalStorage(BaseStorage):
                 if (
                     isinstance(self._backend, BaseJournalLogSnapshot)
                     and study_id != 0
-                    and study_id % self._backend.snapshot_interval == 0
+                    and study_id % SNAPSHOT_INTERVAL == 0
                 ):
                     self._backend.save_snapshot(pickle.dumps(self._replay_result))
                 return study_id
@@ -263,7 +265,7 @@ class JournalStorage(BaseStorage):
         if (
             isinstance(self._backend, BaseJournalLogSnapshot)
             and trial_id != 0
-            and trial_id % self._backend.snapshot_interval == 0
+            and trial_id % SNAPSHOT_INTERVAL == 0
         ):
             self._backend.save_snapshot(pickle.dumps(self._replay_result))
         return trial_id