… tests/storages_tests/test_journal.py