<!-- Thank you for creating a pull request! In general, we merge your pull request after it gets two or more approvals. To proceed to the review process by the maintainers, please make sure that the PR meets the following conditions: (1) it passes all CI checks, and (2) it is neither in the draft nor WIP state. If you wish to discuss the PR in the draft state or need any other help, please mention the Optuna development team in the PR. -->

## Motivation
<!-- Describe your motivation why you will submit this PR. This is useful for reviewers to understand the context of PR. -->
Recently, `JournalStorage` was added to Optuna in https://github.com/optuna/optuna/pull/3854 for file backend log-based storage, and its Redis backend support was added in https://github.com/optuna/optuna/pull/4086. They are ambitious features for distributed optimization in NFS and in environments where RDBs are not readily available. However, there are performance concerns with `JournalStorage` when restarting studies with a large number of trials or when launching a large number of studies. This PR aims to resolve this concern by providing a snapshot feature, especially for `JournalStorage` with Redis as the backend.

## Description of the changes
<!-- Describe the changes in this PR. -->
- Add an abstract class `BaseJournalLogSnapshot` that the log storage concrete class should multiply inherit from in addition to `BaseJournalLogStorage`. This is the same approach as `BaseHeartbeat`. To reduce complexity for developers implementing concrete classes of log storage, `BaseJournalLogSnapshot` requires the implementation of two simple abstract methods. One is `save_snapshot` and the other is `load_snapshot`.
- The `save_snapshot` requires a pickle byte string and should save it in the backend as a snapshot.
- The `load_snapshot` takes a loader function, builds a candidate snapshot internally, and executes that loader function. The loader functions are provided by `JournalStorage` and need not be implemented by developers implementing log storage concrete classes. All he needs to understand is that in `load_snapshot` he extracts the pickled bytes string, which is a candidate snapshot, from the backend and passes it to the loader function.
- Give the concrete implementation above for `JournalRedisStorage`. The reason given for `JournalRedisStorage` is that it is obvious where to store the snapshot.
- Add tests.

## A simple benchmark

```python
import argparse
import time
from typing import Sequence, Tuple

import numpy as np
import redis

import optuna

REDIS_URL = "redis://localhost:6379"


def f(x: float, y: float) -> float:
    return (x - 3) ** 2 + y


def objective(trial: optuna.Trial) -> float:
    x = trial.suggest_float("x", -10, 10)
    y = trial.suggest_float("y", -10, 10)
    trial.report(x, 0)
    trial.report(y, 1)
    trial.set_user_attr("x", x)
    trial.set_system_attr("y", y)
    return f(x, y)

def get_storage(storage_mode: str) -> optuna.storages.BaseStorage:
    if storage_mode == "sqlite":
        return optuna.storages.RDBStorage(SQLITE_URL)
    elif storage_mode == "redis":
       return optuna.storages.RedisStorage(REDIS_URL)
    elif storage_mode == "journal":
        return optuna.storages.JournalStorage(optuna.storages.JournalFileStorage(JOURNAL_FILE))
    elif storage_mode == "journal-redis":
        return optuna.storages.JournalStorage(optuna.storages.JournalRedisStorage(REDIS_URL))
    else:
        assert False, f"Invalid storage mode {storage_mode}"


def test(storage_mode: str):

    n_studies = 30
    n_trials = 300
    study_name = "TEST"

    try:
        optuna.delete_study(storage=storage, study_name=study_name)
    except Exception as e:
        print(e)
    
    start = time.time()
    for _ in range(n_studies):
        storage = get_storage(storage_mode)
        study = optuna.create_study(storage=storage, study_name=study_name, load_if_exists=True, sampler=optuna.samplers.RandomSampler())
        study.optimize(objective, n_trials=n_trials)
    elapsed = time.time() - start
    print(elapsed)

    study = optuna.load_study(study_name=study_name, storage=storage)
    trials = study.trials
    assert len(trials) == n_studies * n_trials


if __name__ == "__main__":
    parser = argparse.ArgumentParser()
    parser.add_argument("storage_mode", type=str)
    args = parser.parse_args()

    r = redis.Redis.from_url(REDIS_URL)
    r.flushdb()
 
    test(args.storage_mode)


```

|  master  |  this PR  |
| ---- | ---- |
|  260.6 sec  |  163.3 sec  |