@wattlebirdaz 2022-11-11
Just a note but we might want a log compaction feature in the future to reduce storage usage because we will not need the old logs once we have the snapshot.